Refactor: move traits; new trait to remove duplications; more gateway defaults.
