+++ src/Message/Response.php
@@ -5,6 +5,7 @@ namespace Omnipay\SagePay\Message;
 use Omnipay\Common\Message\AbstractResponse;
 use Omnipay\Common\Message\RedirectResponseInterface;
 use Omnipay\Common\Message\RequestInterface;
+use Omnipay\SagePay\Traits\ResponseFieldsTrait;
 
 /**
  * Sage Pay Response
